fix: correct singular time unit display in timeAgo() and use absolute dates on detail pages
Cast floor() results to int for proper strict comparison, fixing "1 minutes ago"
displaying instead of "1 minute ago". Also switch detail page templates to show
absolute dates instead of relative dates.

+++ site/plugins/panel-helpers/index.php
@@ -20,31 +20,31 @@ Kirby::plugin('custom/panel-helpers', [
       if ($diff < 60) return 'just now';
 
       if ($diff < 3600) {
-        $minutes = floor($diff / 60);
+        $minutes = (int)floor($diff / 60);
         return $minutes === 1 ? '1 minute ago' : $minutes . ' minutes ago';
       }
 
       if ($diff < 86400) {
-        $hours = floor($diff / 3600);
+        $hours = (int)floor($diff / 3600);
         return $hours === 1 ? '1 hour ago' : $hours . ' hours ago';
       }
 
       if ($diff < 604800) {
-        $days = floor($diff / 86400);
+        $days = (int)floor($diff / 86400);
         return $days === 1 ? '1 day ago' : $days . ' days ago';
       }
 
       if ($diff < 2592000) {
-        $weeks = floor($diff / 604800);
+        $weeks = (int)floor($diff / 604800);
         return $weeks === 1 ? '1 week ago' : $weeks . ' weeks ago';
       }
 
       if ($diff < 31536000) {
-        $months = floor($diff / 2592000);
+        $months = (int)floor($diff / 2592000);
         return $months === 1 ? '1 month ago' : $months . ' months ago';
       }
 
-      $years = floor($diff / 31536000);
+      $years = (int)floor($diff / 31536000);
       return $years === 1 ? '1 year ago' : $years . ' years ago';
     },
